diagnostics: add detailed request & response request and response logging

+++ shared/src/commonMain/kotlin/com/vahitkeskin/loopsweep/domain/usecase/CleanRoomUseCase.kt
@@ -3,7 +3,7 @@ package com.vahitkeskin.loopsweep.domain.usecase
 import com.vahitkeskin.loopsweep.domain.repository.VacuumRepository
 
 class CleanRoomUseCase(private val repository: VacuumRepository) {
-    suspend operator fun invoke(host: String, token: String, roomId: Int, repeats: Int): Result<String> {
+    suspend operator fun invoke(host: String, token: String, roomId: Long, repeats: Int): Result<String> {
         return repository.cleanRoom(host, token, roomId, repeats)
     }
 }
